requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
psycopg==3.2.3
environs[django]==11.2.1
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
djangorestframework==3.15.2